fix(payment): extract PLANS/ANNUAL_PLANS to planConstants.ts
Moving static plan data out of payment.ts means vi.mock("@/services/payment")
no longer suppresses PLANS — any test that mocks the payment canister service
no longer breaks components that only need pricing constants (UpgradeModal,
PricingPage, SettingsPage). payment.ts re-exports everything for backwards
compatibility; getPlan uses a local import from planConstants.
