+++ frontend/src/pages/PricingPage.tsx
@@ -2,8 +2,7 @@ import React, { useState, useEffect } from "react";
 import { Link, useNavigate, useSearchParams } from "react-router-dom";
 import { CheckCircle, Sparkles } from "lucide-react";
 import { Button } from "@/components/Button";
-import { PLANS, ANNUAL_PLANS } from "@/services/payment";
-import type { Plan, PlanTier, BillingCycle } from "@/services/payment";
+import { PLANS, ANNUAL_PLANS, type Plan, type PlanTier, type BillingCycle } from "@/services/planConstants";
 import { useAuth } from "@/contexts/AuthContext";
 import { useAuthStore } from "@/store/authStore";
 import { COLORS, FONTS, RADIUS, SHADOWS } from "@/theme";
