+++ frontend/src/pages/SettingsPage.tsx
@@ -5,7 +5,8 @@ import { Layout } from "@/components/Layout";
 import { Button } from "@/components/Button";
 import { Badge } from "@/components/Badge";
 import { authService } from "@/services/auth";
-import { PLANS, paymentService, type PlanTier } from "@/services/payment";
+import { PLANS, type PlanTier } from "@/services/planConstants";
+import { paymentService } from "@/services/payment";
 import { winBackService } from "@/services/winBackService";
 import { agentProfileService, type AgentProfile } from "@/services/agentProfile";
 import { useAuthStore } from "@/store/authStore";
